// src/shared/state/post/post.action-types.ts
export enum PostActionType {
  FETCH_POSTS_START = 'FETCH_POSTS_START',
  FETCH_POSTS_SUCCESS = 'FETCH_POSTS_SUCCESS',
  FETCH_POSTS_FAILED = 'FETCH_POSTS_FAILED',

  FETCH_SINGLE_POST_START = 'FETCH_SINGLE_POST_START',
  FETCH_SINGLE_POST_SUCCESS = 'FETCH_SINGLE_POST_SUCCESS',
  FETCH_SINGLE_POST_FAILED = 'FETCH_SINGLE_POST_FAILED',

  LIKE_POST_SUCCESS = 'LIKE_POST_SUCCESS',
  LIKE_POST_FAILED = 'LIKE_POST_FAILED',

  UNLIKE_POST_SUCCESS = 'UNLIKE_POST_SUCCESS',
  UNLIKE_POST_FAILED = 'UNLIKE_POST_FAILED',

  COMMENT_POST_START = 'COMMENT_POST_START',
  COMMENT_POST_SUCCESS = 'COMMENT_POST_SUCCESS',
  COMMENT_POST_FAILED = 'COMMENT_POST_FAILED',
}